Move commandMaker to its own module, fix related tests
